Fix Space no longer toggling play/pause after clicking the player button
Clicking the video player's Play/Pause (or Stop / full-screen) button
with the mouse left keyboard focus on it. Avalonia's Button then both
marks Space handled on key-down and raises Click on Space key-up while
focused, so Space started the global play/pause shortcut on press and
the button re-toggled it on release - playback ran only while Space was
held and stopped when released (issue #12759). SE 4 was unaffected
because its transport buttons were not keyboard-focusable.

The waveform toolbar already avoided this with a NonSpaceButton subclass
that swallows Space; the video player toolbar used plain Buttons. Promote
NonSpaceButton to a shared control (src/ui/Controls) and use it for the
player's Play, Stop and full-screen buttons so Space stays owned by the
global shortcut regardless of which transport button has focus.

Fixes #12759
